# api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  models.CreateProvider:
    properties:
      active:
        type: boolean
      name:
        type: string
      phone_number:
        type: string
    type: object
...
  models.Provider:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
    type: object
  models.ProviderGetListResponse:
    properties:
      count:
        type: string
      sale_centers:
        items:
          $ref: '#/definitions/models.Provider'
        type: array
    type: object
...
  models.UpdateProvider:
    properties:
      active:
        type: boolean
      id:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://demo.com/support
  description: Create  Go REST API with JWT Authentication in Gin Framework
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: demo.com
  title: Swagger JWT API
  version: "1.0"
paths:
...
  /provider:
    get:
      consumes:
      - application/json
      description: Get List Provider
      operationId: get_list_provider
      parameters:
      - description: limit
        in: query
        name: limit
        type: number
      - description: offset
        in: query
        name: offset
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: GetListProviderResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ProviderGetListResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List Provider
      tags:
      - Provider
    post:
      consumes:
      - application/json
      description: Create Provider
      operationId: create_provider
      parameters:
      - description: CreateProviderRequestBody
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.CreateProvider'
      produces:
      - application/json
      responses:
        "200":
          description: ProviderBody
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Provider'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Provider
      tags:
      - Provider
  /provider/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Provider
      operationId: delete_provider
      parameters:
      - description: DeleteProviderPath
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deleted Provider
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Provider
      tags:
      - Provider
    get:
      consumes:
      - application/json
      description: Get By Id Provider
      operationId: get_by_id_provider
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ProviderBody
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Provider'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get By Id Provider
      tags:
      - Provider
    put:
      consumes:
      - application/json
      description: Update Provider
      operationId: update_provider
      parameters:
      - description: ProviderPrimaryKey_ID
        in: path
        name: id
        required: true
        type: string
      - description: UpdateBranchBody
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProvider'
      produces:
      - application/json
      responses:
        "200":
          description: Updated Provider
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Provider
      tags:
      - Provider
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"